regen all snaps
